+++ internal/caches/twitchusernamecache/cache_test.go
@@ -1,0 +1,1 @@
+package twitchusernamecache
